Move buildTxQuery into @iov/bns
