+++ packages/iov-tendermint-rpc/src/index.ts
@@ -8,7 +8,6 @@ export {
   BlockResultsRequest,
   BroadcastTxRequest,
   BroadcastTxParams,
-  buildTxQuery,
   CommitRequest,
   GenesisRequest,
   HealthRequest,
@@ -19,7 +18,6 @@ export {
   StatusRequest,
   SubscriptionEventType,
   TxParams,
-  TxQuery,
   TxRequest,
   TxSearchParams,
   TxSearchRequest,
