+++ packages/iov-tendermint-rpc/types/index.d.ts
@@ -1,4 +1,4 @@
 export { Client } from "./client";
-export { AbciInfoRequest, AbciQueryParams, AbciQueryRequest, BlockRequest, BlockchainRequest, BlockResultsRequest, BroadcastTxRequest, BroadcastTxParams, buildTxQuery, CommitRequest, GenesisRequest, HealthRequest, Method, Request, QueryString, QueryTag, StatusRequest, SubscriptionEventType, TxParams, TxQuery, TxRequest, TxSearchParams, TxSearchRequest, ValidatorsRequest, } from "./requests";
+export { AbciInfoRequest, AbciQueryParams, AbciQueryRequest, BlockRequest, BlockchainRequest, BlockResultsRequest, BroadcastTxRequest, BroadcastTxParams, CommitRequest, GenesisRequest, HealthRequest, Method, Request, QueryString, QueryTag, StatusRequest, SubscriptionEventType, TxParams, TxRequest, TxSearchParams, TxSearchRequest, ValidatorsRequest, } from "./requests";
 export * from "./responses";
 export { HttpClient, WebsocketClient } from "./rpcclient";
